requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Brotli==1.0.7
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coverage==5.1
cryptography==2.9.2
cycler==0.10.0
daphne==2.5.0
Django==3.0.5
django-debug-toolbar==2.2
django-ipware==3.0.0
django-mathfilters==1.0.0
django-nose==1.4.6
django-redis==4.12.1
django-tz-detect==0.3.0
djangorestframework==3.11.0
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.5.0
freezegun==0.3.15
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msgpack==0.6.2
nose==1.3.7
numpy==1.18.5
pandas==1.0.4
pandas-datareader==0.8.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
sentry-sdk==0.16.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
